Move Category entity to package directory
