+++ feature/ontology/src/commonMain/kotlin/category/interactors/CreateNewCategoryUseCase.kt
@@ -3,7 +3,7 @@ package org.pointyware.commonsense.feature.ontology.category.interactors
 import org.pointyware.commonsense.core.common.Uuid
 import org.pointyware.commonsense.feature.ontology.category.data.CategoryRepository
 import org.pointyware.commonsense.feature.ontology.data.ConceptEditorController
-import org.pointyware.commonsense.feature.ontology.entities.Category
+import org.pointyware.commonsense.feature.ontology.Category
 
 /**
  * Creates a new concept in the currently selected category.
