+++ feature/ontology/src/commonMain/kotlin/category/interactors/GetSelectedCategoryUseCase.kt
@@ -4,7 +4,7 @@ import org.pointyware.commonsense.core.common.Uuid
 import org.pointyware.commonsense.feature.ontology.Concept
 import org.pointyware.commonsense.feature.ontology.category.data.CategoryRepository
 import org.pointyware.commonsense.feature.ontology.data.ConceptEditorController
-import org.pointyware.commonsense.feature.ontology.entities.Category
+import org.pointyware.commonsense.feature.ontology.Category
 
 /**
  */
